Update PyQt5 to PyQt6 and set application style
The upgrade from PyQt5 to PyQt6 is reflected in the import statements. Additionally, a style is set for the application using the setStyle method, with 'WindowsVista' as the selected style. This can be replaced with any available style. Included as well is a minor tweak in the setWindowIcon function parameters for compatibility.