Update main window setup and UI library
Updated the Main window setup in main_windows.py to use PyQt6 instead of PyQt5. A window icon has also been added to the application for aesthetic appeal. The addition of the 'setWindowIcon' method enables this change. The rest of the changes are just for formatting and do not change the function of the code.