Add automatic update feature to about tab
The implementation of a new feature to the about tab that lets users check and download the latest version of the application. The new functionality involves making a request to the GitHub API to get the latest release, comparing it to the current app version, and offering to download and install the update if a newer version is found. A progress dialog was also added to help inform users of the update process.