Update application renaming logic in settings tab
This commit updates the application renaming logic in the settings tab to accommodate changing the name from "ADB_Controller" to "ADB_Manager". It also adjusts the update notification to display the new name of the application upon successful installation. The path for launching the application post-update has been updated accordingly.